requirements.txt
streamlit
PyPDF2
python-docx
google-generativeai
gtts
